+++ _setup/setup_tools_windows.sh
@@ -4,18 +4,24 @@ cd tools_windows
   git clone --depth=1 --recursive https://github.com/bitsadmin/wesng.git
   git clone --depth=1 --recursive https://github.com/deepzec/Bad-Pdf.git
 
-  wget https://download.sysinternals.com/files/SysinternalsSuite.zip
-  unzip SysinternalsSuite.zip -d sysi
-  rm SysinternalsSuite.zip
+  if [ ! -d sysi ]; then
+    wget https://download.sysinternals.com/files/SysinternalsSuite.zip
+    unzip SysinternalsSuite.zip -d sysi
+    rm SysinternalsSuite.zip
+  fi
 
-  curl -s https://api.github.com/repos/gentilkiwi/mimikatz/releases/latest \
-   | grep "zipball_url.*zip" \
-   | cut -d : -f 2,3 \
-   | tr -d \" \
-   | tr -d , \
-   | wget -qi -O mimikatz_trunk.zip -
-  unzip mimikatz_trunk.zip -d mimikatz
-  rm mimikatz_trunk.zip
+  if [ ! -d mimikatz ]; then
+    wget https://github.com/gentilkiwi/mimikatz/releases/download/2.2.0-20200104/mimikatz_trunk.zip
+    #TODO This fails sometimes due to API rate limiting
+    #curl -s https://api.github.com/repos/gentilkiwi/mimikatz/releases/latest \
+    # | grep "zipball_url.*zip" \
+    # | cut -d : -f 2,3 \
+    # | tr -d \" \
+    # | tr -d , \
+    # | wget -qi -O mimikatz_trunk.zip -
+    unzip mimikatz_trunk.zip -d mimikatz
+    rm mimikatz_trunk.zip
+  fi
 
 cd -
